@@ -22,19 +22,20 @@
   "dependencies": {
     "@plugnet/util": "^0.94.100",
     "@plugnet/wasm-crypto-js": "^0.11.102",
-    "@polkadot/keyring": "^1.0.1",
-    "@polkadot/reactnative-identicon": "^0.42.0-beta.28",
-    "@polkadot/util-crypto": "^1.0.1",
+    "@polkadot/api": "^0.90.0-beta.66",
+    "@polkadot/reactnative-identicon": "^0.42.0-beta.26",
+    "@polkadot/util": "^1.0.1",
+    "@polkadot/util-crypto": "1.1.0-beta.1",
+    "@polkadot/wasm-crypto": "^0.13.1",
     "@react-native-community/netinfo": "^4.1.1",
     "@tradle/react-native-http": "^2.0.1",
     "bignumber.js": "^9.0.0",
     "hoist-non-react-statics": "^3.3.0",
-    "lodash": "^4.17.15",
     "node-libs-react-native": "^1.0.3",
     "prop-types": "^15.6.1",
-    "react": "16.8.6",
+    "react": "^16.9.0",
     "react-native": "0.60.4",
-    "react-native-camera": "^2.11.1",
+    "react-native-camera": "^3.0.0",
     "react-native-gesture-handler": "^1.3.0",
     "react-native-keyboard-aware-scroll-view": "^0.8.0",
     "react-native-markdown-renderer": "^3.2.8",
@@ -61,19 +62,12 @@
     "babel-plugin-rewrite-require": "^1.14.5",
     "eslint": "^6.0.1",
     "jest": "^24.8.0",
-    "jest-react-native": "^18.0.0",
     "jetifier": "^1.6.3",
     "metro-react-native-babel-preset": "^0.55.0",
     "pre-git": "^3.14.0",
     "prettier": "1.18.2",
-    "reactotron-react-native": "^3.6.4",
-    "react-test-renderer": "16.8.6"
-  },
-  "jest": {
-    "preset": "react-native",
-    "setupFiles": [
-      "<rootDir>/jest-setup.js"
-    ]
+    "react-test-renderer": "16.8.6",
+    "reactotron-react-native": "^3.6.4"
   },
   "release": {
     "analyzeCommits": "simple-commit-message"
