@@ -55,7 +55,11 @@ public class EthkeyBridge extends ReactContextBaseJavaModule {
 
     @ReactMethod
     public void brainWalletSign(String seed, String message, Promise promise) {
-        promise.resolve(ethkeyBrainwalletSign(seed, message));
+        try {
+            promise.resolve(ethkeyBrainwalletSign(seed, message));
+        } catch (Exception e) {
+            promise.reject("invalid phrase", "invalid phrase");
+        }
     }
 
     @ReactMethod
@@ -70,7 +74,20 @@ public class EthkeyBridge extends ReactContextBaseJavaModule {
 
     @ReactMethod
     public void keccak(String data, Promise promise) {
-        promise.resolve(ethkeyKeccak(data));
+        try {
+            promise.resolve(ethkeyKeccak(data));
+        } catch (Exception e) {
+            promise.reject("invalid data, expected hex-encoded string", "invalid data, expected hex-encoded string");
+        }
+    }
+
+    @ReactMethod
+    public void blake2s(String data, Promise promise) {
+        try {
+            promise.resolve(ethkeyBlake(data));
+        } catch (Exception e) {
+            promise.reject("invalid data, expected hex-encoded string", "invalid data, expected hex-encoded string");
+        }
     }
 
     @ReactMethod
@@ -111,15 +128,46 @@ public class EthkeyBridge extends ReactContextBaseJavaModule {
         }
     }
 
+    @ReactMethod
+    public void qrCodeHex(String data, Promise promise) {
+        try {
+            promise.resolve(ethkeyQrCodeHex(data));
+        } catch (Exception e) {
+            promise.reject("failed to create QR code", "failed to create QR code");
+        }
+    }
+
+    @ReactMethod
+    public void substrateAddress(String seed, int prefix, Promise promise) {
+        try {
+            promise.resolve(substrateBrainwalletAddress(seed, prefix));
+        } catch (Exception e) {
+            promise.reject("invalid phrase", "invalid phrase");
+        }
+    }
+
+    @ReactMethod
+    public void substrateSign(String seed, String message, Promise promise) {
+        try {
+            promise.resolve(substrateBrainwalletSign(seed, message));
+        } catch (Exception e) {
+            promise.reject("invalid phrase", "invalid phrase");
+        }
+    }
+
     private static native String ethkeyBrainwalletAddress(String seed);
     private static native String ethkeyBrainwalletBIP39Address(String seed);
     private static native String ethkeyBrainwalletSign(String seed, String message);
     private static native String ethkeyRlpItem(String data, int position);
     private static native String ethkeyKeccak(String data);
+    private static native String ethkeyBlake(String data);
     private static native String ethkeyEthSign(String data);
     private static native String ethkeyBlockiesIcon(String seed);
     private static native String ethkeyRandomPhrase();
     private static native String ethkeyEncryptData(String data, String password);
     private static native String ethkeyDecryptData(String data, String password);
     private static native String ethkeyQrCode(String data);
+    private static native String ethkeyQrCodeHex(String data);
+    private static native String substrateBrainwalletAddress(String seed, int prefix);
+    private static native String substrateBrainwalletSign(String seed, String message);
 }
