@@ -32,10 +32,9 @@ public class MainActivity extends ReactActivity {
     @Override
     protected ReactActivityDelegate createReactActivityDelegate() {
         return new ReactActivityDelegate(this, getMainComponentName()) {
-
             @Override
             protected ReactRootView createRootView() {
-                return new RNGestureHandlerEnabledRootView(MainActivity.this);
+            return new RNGestureHandlerEnabledRootView(MainActivity.this);
             }
         };
     }
