@@ -3,17 +3,17 @@ package io.parity.signer;
 import android.app.Application;
 
 import com.facebook.react.ReactApplication;
+import com.reactlibrary.RNSecureStoragePackage;
+import org.reactnative.camera.RNCameraPackage;
+import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 import com.horcrux.svg.SvgPackage;
 import com.bitgo.randombytes.RandomBytesPackage;
 import com.reactnativecommunity.netinfo.NetInfoPackage;
-import org.reactnative.camera.RNCameraPackage;
-import com.reactlibrary.RNSecureStoragePackage;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
 
-import com.swmansion.gesturehandler.react.RNGestureHandlerPackage;
 
 import java.util.Arrays;
 import java.util.List;
@@ -30,11 +30,11 @@ public class MainApplication extends Application implements ReactApplication {
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
           new MainReactPackage(),
-            new SvgPackage(),
-            new RandomBytesPackage(),
-            new NetInfoPackage(),
-            new RNCameraPackage(),
-            new RNSecureStoragePackage(),
+          new RNSecureStoragePackage(),
+          new RNCameraPackage(),
+          new SvgPackage(),
+          new NetInfoPackage(),
+          new RandomBytesPackage(),
           new EthkeyBridgePackage(),
           new RNGestureHandlerPackage()
       );
