@@ -1,4 +1,33 @@
+// Copyright 2015-2019 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+const { defaults } = require('jest-config');
+
 module.exports = {
+  preset: "react-native",
+  moduleFileExtensions: [...defaults.moduleFileExtensions, 'ts', 'tsx'],
   verbose: true,
-  roots: ['<rootDir>/src'],
-};
+  roots: [
+    "<rootDir>/src"
+  ],
+  setupFiles: [
+    "<rootDir>/jest-setup.js"
+  ],
+  testEnvironment: "node",
+  testPathIgnorePatterns: [
+    "/node_modules/"
+  ]
+}
