@@ -15,6 +15,7 @@
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
 mod eth;
+mod sr25519;
 mod util;
 
 use util::StringPtr;
@@ -25,7 +26,7 @@ use rlp::decode_list;
 use rustc_hex::{ToHex, FromHex};
 use tiny_keccak::Keccak;
 use tiny_keccak::keccak256 as keccak;
-
+use blake2_rfc::blake2s::blake2s;
 use std::num::NonZeroU32;
 
 // 10240 is always non-zero, ergo this is safe
@@ -66,6 +67,33 @@ pub unsafe extern fn ethkey_keypair_sign(keypair: *mut KeyPair, message: *mut St
     let message: Vec<u8> = (*message).as_str().from_hex().unwrap();
     let signature = keypair.sign(&message).unwrap().to_hex();
     Box::into_raw(Box::new(signature))
+}
+
+fn qrcode_bytes(data: &[u8]) -> Option<String> {
+    use qrcodegen::{QrCode, QrCodeEcc};
+    use pixelate::{Image, Color, BLACK};
+
+    let qr = QrCode::encode_binary(data, QrCodeEcc::Medium).ok()?;
+
+    let palette = &[Color::Rgba(255,255,255,0), BLACK];
+    let mut pixels = Vec::with_capacity((qr.size() * qr.size()) as usize);
+
+    for y in 0..qr.size() {
+        for x in 0..qr.size() {
+            pixels.push(qr.get_module(x, y) as u8);
+        }
+    }
+
+    let mut result = Vec::new();
+
+    Image {
+        palette,
+        pixels: &pixels,
+        width: qr.size() as usize,
+        scale: 16,
+    }.render(&mut result).ok()?;
+
+    Some(base64png(&result))
 }
 
 export! {
@@ -112,6 +140,13 @@ export! {
         let data: Vec<u8> = data.from_hex().ok()?;
 
         Some(keccak(&data).to_hex())
+    }
+
+    @Java_io_parity_signer_EthkeyBridge_ethkeyBlake
+    fn blake(data: &str) -> Option<String> {
+        let data: Vec<u8> = data.from_hex().ok()?;
+
+        Some(blake2s(32, &[], &data).as_bytes().to_hex())
     }
 
     @Java_io_parity_signer_EthkeyBridge_ethkeyBlockiesIcon
@@ -169,36 +204,35 @@ export! {
 
     @Java_io_parity_signer_EthkeyBridge_ethkeyQrCode
     fn qrcode(data: &str) -> Option<String> {
-        use qrcodegen::{QrCode, QrCodeEcc};
-        use pixelate::{Image, Color, BLACK};
+        qrcode_bytes(data.as_bytes())
+    }
 
-        let qr = QrCode::encode_binary(data.as_bytes(), QrCodeEcc::Medium).ok()?;
+    @Java_io_parity_signer_EthkeyBridge_ethkeyQrCodeHex
+    fn qrcode_hex(data: &str) -> Option<String> {
+        qrcode_bytes(&data.from_hex::<Vec<u8>>().ok()?)
+    }
 
-        let palette = &[Color::Rgba(255,255,255,0), BLACK];
-        let mut pixels = Vec::with_capacity((qr.size() * qr.size()) as usize);
+    @Java_io_parity_signer_EthkeyBridge_substrateBrainwalletAddress
+    fn substrate_brainwallet_address(suri: &str, prefix: u8) -> Option<String> {
+        let keypair = sr25519::KeyPair::from_suri(suri)?;
 
-        for y in 0..qr.size() {
-            for x in 0..qr.size() {
-                pixels.push(qr.get_module(x, y) as u8);
-            }
-        }
+        Some(keypair.ss58_address(prefix))
+    }
 
-        let mut result = Vec::new();
+    @Java_io_parity_signer_EthkeyBridge_substrateBrainwalletSign
+    fn substrate_brainwallet_sign(suri: &str, message: &str) -> Option<String> {
+        let keypair = sr25519::KeyPair::from_suri(suri)?;
 
-        Image {
-            palette,
-            pixels: &pixels,
-            width: qr.size() as usize,
-            scale: 16,
-        }.render(&mut result).ok()?;
+        let message: Vec<u8> = message.from_hex().ok()?;
+        let signature = keypair.sign(&message);
 
-        Some(base64png(&result))
+        Some(signature.to_hex())
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::rlp_item;
+    use super::*;
 
     #[test]
     fn test_rlp_item() {
@@ -209,5 +243,25 @@ mod tests {
         assert_eq!(rlp_item(rlp, 3), Some("095e7baea6a6c7c4c2dfeb977efac326af552d87".into()));
         assert_eq!(rlp_item(rlp, 4), Some("0a".into()));
         assert_eq!(rlp_item(rlp, 5), Some("".into()));
+    }
+
+    #[test]
+    fn test_substrate_brainwallet_address() {
+        let suri = "grant jaguar wish bench exact find voice habit tank pony state salmon";
+        // Secret seed: 0xb139e4050f80172b44957ef9d1755ef5c96c296d63b8a2b50025bf477bd95224
+        // Public key (hex): 0x944eeb240615f4a94f673f240a256584ba178e22dd7b67503a753968e2f95761
+        let expected = "5FRAPSnpgmnXAnmPVv68fT6o7ntTvaZmkTED8jDttnXs9k4n";
+        let generated = substrate_brainwallet_address(suri, 42).unwrap();
+
+        assert_eq!(expected, generated);
+    }
+
+    #[test]
+    fn test_substrate_brainwallet_address_suri() {
+        let suri = "grant jaguar wish bench exact find voice habit tank pony state salmon//hard/soft/0";
+        let expected = "5D4kaJXj5HVoBw2tFFsDj56BjZdPhXKxgGxZuKk4K3bKqHZ6";
+        let generated = substrate_brainwallet_address(suri, 42).unwrap();
+
+        assert_eq!(expected, generated);
     }
 }
