@@ -16,59 +16,55 @@
 
 'use strict';
 
+import Identicon from '@polkadot/reactnative-identicon';
 import PropTypes from 'prop-types';
-import React from 'react';
-import { Image, StyleSheet, View } from 'react-native';
+import React, { useEffect, useState } from 'react';
+import { Image } from 'react-native';
+
+import { NetworkProtocols } from '../constants'
 import { blockiesIcon } from '../util/native';
 
-export default class AccountIcon extends React.PureComponent {
+export default function AccountIcon (props) {
 
-  constructor(...args) {
-    super(...args);
-    this.displayIcon = this.displayIcon.bind(this);
-  }
-
-  static propTypes = {
-    seed: PropTypes.string.isRequired
+  AccountIcon.propTypes = {
+    address: PropTypes.string.isRequired,
+    protocol: PropTypes.string.isRequired
   };
 
-  state = {};
+  const {address, protocol, style} = props;
+  const [ethereumIconUri, setEthereumIconUri] = useState('');
 
-  async displayIcon(seed) {
-    try {
-      let icon = await blockiesIcon(seed);
-      this.setState({
-        icon: icon
-      });
-    } catch (e) {
-      console.log(e);
+  useEffect(() => {
+    if (protocol === NetworkProtocols.ETHEREUM) {
+      loadEthereumIcon(address);
     }
+  },[protocol, address])
+
+  const loadEthereumIcon = function (address){
+    blockiesIcon('0x'+address)
+    .then((ethereumIconUri) => {
+      setEthereumIconUri(ethereumIconUri);
+    })
+    .catch(console.error)
   }
 
-  componentDidMount() {
-    this.displayIcon(this.props.seed);
-  }
+  if (protocol === NetworkProtocols.SUBSTRATE) {
 
-  componentWillReceiveProps(newProps) {
-    if (newProps.seed !== this.props.seed) {
-      this.displayIcon(newProps.seed);
-    }
-  }
-
-  render() {
     return (
-      <View style={styles.identicon}>
-        <Image
-          style={this.props.style || {}}
-          source={{ uri: this.state.icon }}
-        />
-      </View>
+      <Identicon
+        value={address}
+        size={style.width || 50 }
+      />
+    );
+  } else if (protocol === NetworkProtocols.ETHEREUM && ethereumIconUri){
+
+    return (
+      <Image 
+        source={{ uri: ethereumIconUri }} 
+        style={style || { width: 47, height: 47 }}
+      />
     );
   }
-}
 
-const styles = StyleSheet.create({
-  identicon: {
-    alignItems: 'center'
-  }
-});
+  return null;
+}
