@@ -24,9 +24,11 @@ import {
 
 import colors from '../colors';
 import fonts from "../fonts";
+import {NetworkProtocols} from '../constants'
 
 export default function Address (props) {
-  const {address, short = false ,style = {}} = props;
+  const {address, protocol = NetworkProtocols.SUBSTRATE, short = false ,style = {}} = props;
+  const prefix = protocol === NetworkProtocols.ETHEREUM ? '0x' : '';
   let result = address;
 
   if (short) {
@@ -35,7 +37,7 @@ export default function Address (props) {
 
   return (
       <Text numberOfLines={1} style={[style, styles.secondaryText]}>
-        0x{result}
+        {prefix}{result}
       </Text>
   );
 }
