@@ -19,7 +19,7 @@
 import React from 'react';
 import { Linking, ScrollView, StyleSheet, Text, View } from 'react-native';
 import colors from '../colors';
-import fonts from "../fonts";
+import fonts from '../fonts';
 import packageJson from '../../package.json';
 
 export default class About extends React.PureComponent {
@@ -31,7 +31,7 @@ export default class About extends React.PureComponent {
   render() {
     return (
       <ScrollView style={styles.body} contentContainerStyle={{ padding: 20 }}>
-        <Text style={styles.title}>PARITY SIGNER ({packageJson.version})</Text>
+        <Text style={styles.title}>PARITY SIGNER  ({packageJson.version})</Text>
         <View>
           <Text style={styles.text}>
             The Parity Signer mobile application is a secure air-gapped wallet
