@@ -105,6 +105,7 @@ This account can only be recovered with its associated recovery phrase.`,
 
   render() {
     const account = this.props.accounts.getSelected();
+
     if (!account) {
       return null;
     }
@@ -134,7 +135,7 @@ This account can only be recovered with its associated recovery phrase.`,
           title={account.name}
         />
         <View style={styles.qr}>
-          <QrView text={accountId(account)} />
+          <QrView data={accountId(account)} />
         </View>
       </ScrollView>
     );
