@@ -17,12 +17,11 @@
 'use strict';
 
 import React from 'react';
-import { Clipboard, ScrollView, StyleSheet, Text } from 'react-native';
+import { ScrollView, StyleSheet, Text } from 'react-native';
 import { Subscribe } from 'unstated';
 import colors from '../colors';
 import fonts from "../fonts";
 import AccountCard from '../components/AccountCard';
-import Button from '../components/Button';
 import TextInput from '../components/TextInput';
 import AccountsStore from '../stores/AccountsStore';
 
@@ -41,9 +40,11 @@ export default class AccountEdit extends React.PureComponent {
       <Subscribe to={[AccountsStore]}>
         {accounts => {
           const selected = accounts.getSelected();
+
           if (!selected) {
             return null;
           }
+
           return (
             <ScrollView
               style={styles.body}
@@ -54,15 +55,14 @@ export default class AccountEdit extends React.PureComponent {
                 title={selected.name}
                 address={selected.address}
                 networkKey={selected.networkKey}
-                onPress={async () => {
-                  await Clipboard.setString('0x' + selected.address);
-                }}
               />
               <Text style={styles.title}>ACCOUNT NAME</Text>
               <TextInput
                 style={{ marginBottom: 40 }}
-                onChangeText={name => accounts.updateSelected({ name })}
-                onEndEditing={text => accounts.saveSelected()}
+                onChangeText={async (name) => {
+                  accounts.updateSelected({ name });
+                  await accounts.save(accounts.getSelected())
+                }}
                 value={selected.name}
                 placeholder="New name"
               />
