@@ -36,6 +36,7 @@ export default class AccountList extends React.PureComponent {
   };
 
   render() {
+
     return (
       <Subscribe to={[AccountsStore]}>
         {accounts => {
@@ -110,8 +111,9 @@ class AccountListView extends React.PureComponent {
   };
 
   render() {
-    const hasNoAccount = this.props.accounts.length < 1;
-    const { navigate } = this.props.navigation;
+    const { accounts, navigation, onAccountSelected } = this.props;
+    const hasNoAccount = accounts.length < 1;
+    const { navigate } = navigation;
 
     return (
       <View style={styles.body}>
@@ -136,19 +138,20 @@ class AccountListView extends React.PureComponent {
             this.list = list;
           }}
           style={styles.content}
-          data={this.props.accounts}
+          data={accounts}
           keyExtractor={account => accountId(account)}
           ItemSeparatorComponent={() => <View style={{ height: 20 }} />}
           renderItem={({ item: account }) => {
             return (
               <AccountCard
-                title={account.name}
-                style={{ paddingBottom: null }}
                 address={account.address}
                 networkKey={account.networkKey}
                 onPress={() => {
-                  this.props.onAccountSelected(account);
+                  onAccountSelected(account);
                 }}
+                shortAddress
+                style={{ paddingBottom: null }}
+                title={account.name}
               />
             );
           }}
