@@ -1,28 +1,36 @@
-import { EthereumNetworkIds, NetworkTypes } from '../constants';
+import { NetworkProtocols, NETWORK_LIST, SubstrateNetworkKeys } from '../constants';
 
 export function accountId({
   address,
-  networkType = 'ethereum',
-  networkKey = EthereumNetworkIds.FRONTIER
+  networkKey
 }) {
-  if (typeof address !== 'string' || address.length === 0) {
-    throw new Error(`Couldn't create an accountId, missing address`);
+
+  const { ethereumChainId, protocol, genesisHash } = NETWORK_LIST[networkKey];
+
+  if (typeof address !== 'string' || address.length === 0 || !networkKey) {
+    throw new Error(`Couldn't create an accountId, address or networkKey missing`);
   }
-  return `${networkType}:0x${address.toLowerCase()}@${networkKey}`;
+
+  if (protocol === NetworkProtocols.SUBSTRATE){ 
+    return `${protocol}:${address}:${genesisHash}`;
+  } else {
+    return `${protocol}:${address.toLowerCase()}@${ethereumChainId}`;
+  }
 }
 
 export function empty(account = {}) {
   return {
-    name: '',
-    networkType: NetworkTypes.ETHEREUM,
-    networkKey: EthereumNetworkIds.FRONTIER,
-    seed: '',
-    // address for an empty seed phrase
-    address: '00a329c0648769A73afAc7F9381E08FB43dBEA72',
-    createdAt: new Date().getTime(),
-    updatedAt: new Date().getTime(),
+    address: '',
     archived: false,
+    createdAt: new Date().getTime(),
+    derivationPassword: '', 
+    derivationPath:'',
     encryptedSeed: null,
+    name: '',
+    networkKey: SubstrateNetworkKeys.KUSAMA,
+    seed: '',
+    seedPhrase: '',
+    updatedAt: new Date().getTime(),
     validBip39Seed: false,
     ...account
   };
@@ -36,6 +44,7 @@ export function validateSeed(seed, validBip39Seed) {
       valid: false
     };
   }
+
   const words = seed.split(' ');
 
   for (let word of words) {
@@ -53,7 +62,6 @@ export function validateSeed(seed, validBip39Seed) {
       accountRecoveryAllowed: true,
       reason: `This recovery phrase will be treated as a legacy Parity brain wallet.`,
       valid: false
-      
     };
   }
 
